refactor: Request 와 관련된 객체들을 request 패키지로 이동
