+++ src/main/java/webserver/request/exception/IllegalHttpMethodException.java
@@ -1,4 +1,4 @@
-package exception;
+package webserver.request.exception;
 
 public class IllegalHttpMethodException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 HTTP Method 입니다.";
