+++ src/main/java/webserver/request/exception/IllegalProtocolException.java
@@ -1,4 +1,4 @@
-package exception;
+package webserver.request.exception;
 
 public class IllegalProtocolException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 Protocol 입니다.";
