+++ src/main/java/webserver/request/exception/IllegalQueryStringException.java
@@ -1,4 +1,4 @@
-package exception;
+package webserver.request.exception;
 
 public class IllegalQueryStringException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 QueryString 입니다.";
