+++ src/main/java/webserver/request/exception/IllegalQueryStringKeyException.java
@@ -1,4 +1,4 @@
-package exception;
+package webserver.request.exception;
 
 public class IllegalQueryStringKeyException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 QueryString 의 키입니다.";
