+++ src/main/java/webserver/request/exception/IllegalRequestLineException.java
@@ -1,4 +1,4 @@
-package exception;
+package webserver.request.exception;
 
 public class IllegalRequestLineException extends RuntimeException {
     private static final String MESSAGE = "%s; 부적절한 RequestLine 입니다.";
